requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.3
colorama==0.4.6
filelock==3.19.1
fsspec==2025.9.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.45.1
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==2.3.3
openai-whisper==20250625
regex==2025.9.18
requests==2.32.5
sympy==1.14.0
tiktoken==0.11.0
torch==2.8.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
